backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.8.2
pydantic-settings==2.4.0
SQLAlchemy==2.0.32
psycopg2-binary==2.9.9
alembic==1.13.2
redis==5.0.8
python-telegram-bot==21.4
dropbox==12.0.2
python-jose==3.3.0
passlib[bcrypt]==1.7.4
httpx==0.27.2
websockets==12.0
jinja2==3.1.4
pandas==2.2.2
openpyxl==3.1.5
python-Levenshtein==0.25.1
rapidfuzz==3.9.4
orjson==3.10.7
loguru==0.7.2
PyJWT==2.9.0
aiofiles==24.1.0
email-validator==2.1.0.post1

